+++ src/reading.rs
@@ -1,4 +1,7 @@
 //! Describes a sensor reading.
+//!
+//! This is a key data structure. All the data is stored as a single list of all readings from all services.
+
 use crate::value::Value;
 use chrono::prelude::*;
 
@@ -6,20 +9,18 @@ use chrono::prelude::*;
 #[derive(Debug)]
 pub struct Reading {
     /// A sensor. For example: `buienradar::6240::wind_speed_bft`.
+    ///
+    /// Note that sensors do not exist as separate entities. Sensor is only a sort of "label"
+    /// that corresponds to a sensor in the physical world and used to distinguish readings
+    /// between different "real" sensors.
     pub sensor: String,
+
     /// An attached typed value.
     pub value: Value,
-    /// Timestamp when the value was actually measured.
-    pub timestamp: DateTime<Local>,
-}
 
-impl Reading {
-    /// Create a new reading.
-    pub fn new(sensor: String, value: Value, timestamp: Option<DateTime<Local>>) -> Self {
-        Reading {
-            sensor,
-            value,
-            timestamp: timestamp.unwrap_or_else(Local::now),
-        }
-    }
+    /// Timestamp when the value was actually measured. This may be earlier than moment of emitting a reading.
+    pub timestamp: DateTime<Local>,
+
+    /// Should the reading be persisted in the database.
+    pub is_persisted: bool,
 }